+++ src/RampConstAcceleration.h
@@ -3,33 +3,6 @@
 
 #include "common.h"
 
-#if (TICKS_PER_S == 16000000L)
-#define PMF_TICKS_PER_S PMF_CONST_16E6
-#define PMF_TICKS_PER_S_DIV_SQRT_OF_2 PMF_CONST_16E6_DIV_SQRT_OF_2
-#define PMF_ACCEL_FACTOR PMF_CONST_128E12
-#define US_TO_TICKS(u32) (u32 * 16)
-#define TICKS_TO_US(u32) (u32 / 16)
-#elif (TICKS_PER_S == 21000000L)
-#define PMF_TICKS_PER_S PMF_CONST_21E6
-#define PMF_TICKS_PER_S_DIV_SQRT_OF_2 PMF_CONST_21E6_DIV_SQRT_OF_2
-#define PMF_ACCEL_FACTOR PMF_CONST_2205E11
-#define US_TO_TICKS(u32) (u32 * 21)
-#define TICKS_TO_US(u32) (u32 / 21)
-#else
-#define SUPPORT_PMF_TIMER_FREQ_VARIABLES
-#define PMF_TICKS_PER_S pmfl_timer_freq
-#define PMF_TICKS_PER_S_DIV_SQRT_OF_2 pmfl_timer_freq_div_sqrt_of_2
-#define PMF_ACCEL_FACTOR pmfl_timer_freq_square_div_2
-// This overflows for approx. 1s at 40 MHz, only
-#define US_TO_TICKS(u32) \
-  ((uint32_t)((((uint32_t)((u32) * (TICKS_PER_S / 10000L))) / 100L)))
-
-// This calculation needs more work
-#define TICKS_TO_US(u32) \
-  ((uint32_t)((((uint32_t)((u32) / (TICKS_PER_S / 1000000L))) / 1L)))
-
-#endif
-
 struct ramp_ro_s {
   struct ramp_config_s config;
   int32_t target_pos;