+++ CHANGELOG.md
@@ -5,7 +5,7 @@ TODO:
 
 pre-0.29.1:
 - improve rounding for log2/pow2 conversion
-- avr: reduce max speed for 1 stepper to 60kSteps/s from 70kSteps/s
+- avr: reduce max speed for 1 stepper to 50kSteps/s from 70kSteps/s
 
 0.29.0:
 - replace former PoorManFloat implementation `ump_float` being exponent+mantissa with `pmfl_logarithmic` as log2 representation